commons: use long to store updates count of a gauge
to prevent integer overflow

Ticket: #8158
Acked-by: Gerd Behrmann
Target: master
Require-book: no
Require-notes: no
(cherry picked from commit d5512f62eec2ab040df0bbd87e1280c68594ac1f)

+++ modules/common/src/main/java/org/dcache/commons/stats/RequestExecutionTimeGauge.java
@@ -71,7 +71,7 @@ public interface RequestExecutionTimeGauge {
     /**
      * @return the updateNum
      */
-    int getUpdateNum();
+    long getUpdateNum();
 
     /**
      *
